backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.4.0
pydantic==2.11.7
requests==2.32.4
aiohttp==3.9.1
google-adk==1.8.0
google-genai==1.27.0
